updating scripts after tests on pt_br

+++ download.sh
@@ -15,11 +15,17 @@ mkdir -p OpenSubtitles/xml
 mkdir -p OpenSubtitles/"$pairname"/parsed
 mkdir -p OpenSubtitles/"$pairname"/cxt_dataset
 
-wget "http://opus.nlpl.eu/download.php?f=OpenSubtitles/v2018/xml/$src.zip" -O ./OpenSubtitles/xml/"$src".zip
-wget "http://opus.nlpl.eu/download.php?f=OpenSubtitles/v2018/xml/$tgt.zip" -O ./OpenSubtitles/xml/"$tgt".zip
-wget "http://opus.nlpl.eu/download.php?f=OpenSubtitles/v2018/xml/$pairname.xml.gz" -O ./OpenSubtitles/"$pairname"/"$pairname".xml.gz
+if [ ! -f ./OpenSubtitles/xml/"$src".zip ]; then
+  wget "http://opus.nlpl.eu/download.php?f=OpenSubtitles/v2018/xml/$src.zip" -O ./OpenSubtitles/xml/"$src".zip
+  unzip OpenSubtitles/xml/"$src".zip
+fi
 
-unzip OpenSubtitles/xml/"$src".zip
-unzip OpenSubtitles/xml/"$tgt".zip
+if [ ! -f ./OpenSubtitles/xml/"$tgt".zip ]; then
+  wget "http://opus.nlpl.eu/download.php?f=OpenSubtitles/v2018/xml/$tgt.zip" -O ./OpenSubtitles/xml/"$tgt".zip
+  unzip OpenSubtitles/xml/"$tgt".zip
+fi
 
-gunzip OpenSubtitles/"$pairname"/"$pairname".xml.gz
+if [ ! -f ./OpenSubtitles/"$pairname"/"$pairname".xml.gz ]; then
+  wget "http://opus.nlpl.eu/download.php?f=OpenSubtitles/v2018/xml/$pairname.xml.gz" -O ./OpenSubtitles/"$pairname"/"$pairname".xml.gz
+  gunzip OpenSubtitles/"$pairname"/"$pairname".xml.gz
+fi
